Reduce circular dependencies in esperanza
Moves ExtractValidatorPubKey and ExtractValidatorAddress to the esperanza/script.cpp

```diff
< Circular dependency: esperanza/checks -> esperanza/finalizationstate -> esperanza/checks
9d7
< Circular dependency: esperanza/finalizationstate -> validation -> esperanza/finalizationstate
42,43d39
< Circular dependency: consensus/tx_verify -> esperanza/finalizationstate -> validation -> consensus/tx_verify
< Circular dependency: esperanza/finalizationstate -> validation -> finalization/vote_recorder -> esperanza/finalizationstate
```
