+++ src/esperanza/checks.cpp
@@ -6,6 +6,7 @@
 #include <esperanza/adminparams.h>
 #include <esperanza/checks.h>
 #include <esperanza/finalizationstate.h>
+#include <esperanza/script.h>
 #include <script/interpreter.h>
 #include <script/sign.h>
 #include <script/standard.h>
@@ -519,57 +520,6 @@ bool ContextualCheckAdminTx(const CTransaction &tx, CValidationState &err_state,
   }
 
   return true;
-}
-
-bool ExtractValidatorPubkey(const CTransaction &tx, CPubKey &pubkeyOut) {
-  if (tx.IsVote()) {
-
-    std::vector<std::vector<unsigned char>> vSolutions;
-    txnouttype typeRet;
-
-    if (Solver(tx.vout[0].scriptPubKey, typeRet, vSolutions)) {
-      pubkeyOut = CPubKey(vSolutions[0]);
-      return true;
-    }
-  }
-  return false;
-}
-
-bool ExtractValidatorAddress(const CTransaction &tx,
-                             uint160 &validatorAddressOut) {
-
-  switch (tx.GetType()) {
-    case TxType::DEPOSIT:
-    case TxType::LOGOUT: {
-      std::vector<std::vector<unsigned char>> vSolutions;
-      txnouttype typeRet;
-
-      if (Solver(tx.vout[0].scriptPubKey, typeRet, vSolutions)) {
-        assert(typeRet == TX_COMMIT);
-        validatorAddressOut = CPubKey(vSolutions[0]).GetID();
-        return true;
-      }
-      return false;
-    }
-    case TxType::WITHDRAW: {
-
-      const CScript scriptSig = tx.vin[0].scriptSig;
-      auto pc = scriptSig.begin();
-      std::vector<unsigned char> vData;
-      opcodetype opcode;
-
-      // Skip the first value (signature)
-      scriptSig.GetOp(pc, opcode);
-
-      // Retrieve the public key
-      scriptSig.GetOp(pc, opcode, vData);
-      validatorAddressOut = CPubKey(vData).GetID();
-      return true;
-    }
-    default: {
-      return false;
-    }
-  }
 }
 
 }  // namespace esperanza
